detector: add a /metrics endpoint
The /metrics endpoint can be used to get metrics in the prometheus
format. The only metrics exported are about the Go runtime, and we can
add metrics relevant to the application logic in a follow up change.

The metrics can be protected behind authentication if needed, similar to
the health check endpoint.

Metrics to the aggregator can be added in a follow up change.