feat(detector): capture the time of the most recent health check run (#48)
The endpoint `/v1/nodehealth/` exposes the most recent status for all
the health checks the detector is configured to run.

However, there is no information about when it was run. When
troubleshooting a specific health check, it is useful to know if the
information exposes by the endpoint is the most recent one, or if the
check is failing and is not updating the status.

To avoid having to set manually `LastRun` every time we need to update
the struct, a new method named `Update` is added to the type
`HealthCheck`. The signature takes the result and message, and set the
fields on the struct and set `LastRun` too.

The exposed data looks like this:
```
$ curl -s localhost:8083/v1/nodehealth/| jq '.[] | select(.type == "CPUUnderPressure")'
{
  "type": "CPUUnderPressure",
  "result": "false",
  "message": "CPU usage: 0.000000 %",
  "last_run": "2022-07-08T22:37:28.680344271Z"
}
```