feat: add a mechanism to watch for problems in the logs
So far we've focused on reporting problems from external health checks
or by relying on some host level metrics. This change introduces a
different source for problems: logs.

A configuration file (or multiple ones) can be provided to the detector,
with a source (for now only journald) and a list of rules. A log watcher
is then created at startup, and events in the logs are analyzed as they
come and when a log event matches one of the rule, a counter for the
problem associated with the rule is increased.