ci: add golangci-lint to the GitHub action (#54)
golangci-lint is a linter for the Go programming language. It runs many
tools and linter, for example staticcheck, which, using static analysis,
finds bugs and performance issues, offers simplifications, and enforces
style rules.

Making it part of the CI path with help with enforcing good standards
and follow the community style rules.

staticcheck is also added to the Vagrant configuration so it can be run
locally before creating a pull request.