feat(detector): add flag to configure the path to the metrics endpoint
The detector has a metric endpoint that is available at `/v1/metrics/`.
The path is prefixed by `/v1` to be consistent with the other
endpoints (`nodehealth` and `health`). It can also be useful to be
versioned in the case we decide to change its behavior in the future.

However, it's also true that many automatic scrapers expects the metrics
to be available under `/metrics` (which is what the aggregator uses).

This change add a configuration flag to the detector to let the operator
decide which path makes more sense: by default we keep `/v1/metrics/`
but it's now possible to start the detector with
`--prometheus-metrics-path /metrics`.